Fix buildbot container for the use on Linux

+++ .docker/buildbot.Dockerfile
@@ -32,9 +32,10 @@ RUN \
         apt-get update && \
         apt-get install -y devkitpro-pacman && \
     rm -rf /var/lib/apt/lists/* && \
-    { ln -s /proc/self/mounts /etc/mtab || true } && \
+    [ ! -f /etc/mtab ] hotfix=1 \
+    [ -z ${hotfix+z} ] && ln -s /proc/self/mounts /etc/mtab || true && \
     dkp-pacman -S --noconfirm gba-dev && \
-    rm /etc/mtab
+    [ -z ${hotfix+z} ] && rm /etc/mtab || true
 
 RUN \
         git clone https://github.com/ipatix/wav2agb.git && \
